refactor et simplification

+++ src/main/java/project/repositories/CommentRepository.java
@@ -2,8 +2,8 @@ package project.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import project.models.Comment;
+import project.models.entities.Comment;
 
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
+public interface CommentRepository extends JpaRepository<Comment, Long> {
 
 }
