+++ src/main/java/project/repositories/PhotoRepository.java
@@ -4,9 +4,11 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import project.models.Photo;
-import project.models.Room;
+import project.models.entities.Photo;
+import project.models.entities.Room;
 
-public interface PhotoRepository extends JpaRepository<Photo, Integer> {
+public interface PhotoRepository extends JpaRepository<Photo, Long> {
+	
 	public List<Photo> findByRoom(Room room);
+	
 }
